# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Rumi Building Service API",
        "description": "REST API of the Building Management bounded context: buildings, their IoT sensors and resident invitations.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/buildings": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "List buildings",
                "description": "Returns the registered buildings ordered by name (US04). When administratorUserId is sent, only the buildings of that administrator are returned. The list is empty when nothing matches.",
                "operationId": "listBuildings",
                "parameters": [
                    {
                        "name": "administratorUserId",
                        "in": "query",
                        "description": "Returns only the buildings registered by this administrator user",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "c1d2e3f4-a5b6-4c7d-8e9f-0a1b2c3d4e5f"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buildings found (possibly none)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BuildingResponse"
                                    }
                                },
                                "examples": {
                                    "One building": {
                                        "description": "One building",
                                        "value": [
                                            {
                                                "id": "7a9b3c1d-2e4f-4b6a-8c0d-1e2f3a4b5c6d",
                                                "name": "Torre Miraflores",
                                                "address": "Av. Larco 1234, Miraflores",
                                                "floors": 12,
                                                "constructionYear": 2015,
                                                "units": 48,
                                                "status": "ACTIVE",
                                                "administratorUserId": "c1d2e3f4-a5b6-4c7d-8e9f-0a1b2c3d4e5f"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "administratorUserId is not a valid UUID",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                },
                                "examples": {
                                    "Invalid administrator id": {
                                        "description": "Invalid administrator id",
                                        "value": {
                                            "type": "about:blank",
                                            "title": "Bad Request",
                                            "status": 400,
                                            "detail": "Failed to convert 'administratorUserId' with value: 'not-a-uuid'",
                                            "instance": "/api/v1/buildings"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Register a building",
                "description": "Registers a new building to be monitored (US04). The service generates the building id and sets the status to PENDING_SENSORS; the building becomes ACTIVE when its first sensor is activated. The administrator is sent in the body until the IAM service exists.",
                "operationId": "registerBuilding",
                "requestBody": {
                    "description": "Data of the building to register",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BuildingRequest"
                            },
                            "examples": {
                                "Residential tower": {
                                    "description": "Residential tower",
                                    "value": {
                                        "name": "Torre Miraflores",
                                        "address": "Av. Larco 1234, Miraflores",
                                        "floors": 12,
                                        "constructionYear": 2015,
                                        "units": 48,
                                        "administratorUserId": "c1d2e3f4-a5b6-4c7d-8e9f-0a1b2c3d4e5f"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Building registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BuildingResponse"
                                },
                                "examples": {
                                    "Registered building": {
                                        "description": "Registered building",
                                        "value": {
                                            "id": "7a9b3c1d-2e4f-4b6a-8c0d-1e2f3a4b5c6d",
                                            "name": "Torre Miraflores",
                                            "address": "Av. Larco 1234, Miraflores",
                                            "floors": 12,
                                            "constructionYear": 2015,
                                            "units": 48,
                                            "status": "PENDING_SENSORS",
                                            "administratorUserId": "c1d2e3f4-a5b6-4c7d-8e9f-0a1b2c3d4e5f"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "A field is missing, blank or out of range, or the body is not valid JSON",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                },
                                "examples": {
                                    "Invalid fields": {
                                        "description": "Invalid fields",
                                        "value": {
                                            "type": "about:blank",
                                            "title": "Bad Request",
                                            "status": 400,
                                            "detail": "Invalid request content.",
                                            "instance": "/api/v1/buildings",
                                            "errors": {
                                                "floors": "must be greater than or equal to 1",
                                                "name": "must not be blank"
                                            }
                                        }
                                    },
                                    "Malformed body": {
                                        "description": "Malformed body",
                                        "value": {
                                            "type": "about:blank",
                                            "title": "Bad Request",
                                            "status": 400,
                                            "detail": "Failed to read request",
                                            "instance": "/api/v1/buildings"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/buildings/{buildingId}": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Get a building",
                "description": "Returns one building by its identifier, including its current status (US04).",
                "operationId": "getBuilding",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "Identifier of the building",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "7a9b3c1d-2e4f-4b6a-8c0d-1e2f3a4b5c6d"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Building found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BuildingResponse"
                                },
                                "examples": {
                                    "Building": {
                                        "description": "Building",
                                        "value": {
                                            "id": "7a9b3c1d-2e4f-4b6a-8c0d-1e2f3a4b5c6d",
                                            "name": "Torre Miraflores",
                                            "address": "Av. Larco 1234, Miraflores",
                                            "floors": 12,
                                            "constructionYear": 2015,
                                            "units": 48,
                                            "status": "PENDING_SENSORS",
                                            "administratorUserId": "c1d2e3f4-a5b6-4c7d-8e9f-0a1b2c3d4e5f"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "buildingId is not a valid UUID",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                },
                                "examples": {
                                    "Invalid building id": {
                                        "description": "Invalid building id",
                                        "value": {
                                            "type": "about:blank",
                                            "title": "Bad Request",
                                            "status": 400,
                                            "detail": "Failed to convert 'buildingId' with value: 'not-a-uuid'",
                                            "instance": "/api/v1/buildings/not-a-uuid"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No building has that identifier",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                },
                                "examples": {
                                    "Unknown building": {
                                        "description": "Unknown building",
                                        "value": {
                                            "type": "about:blank",
                                            "title": "Not Found",
                                            "status": 404,
                                            "detail": "Building 7a9b3c1d-2e4f-4b6a-8c0d-1e2f3a4b5c6d was not found",
                                            "instance": "/api/v1/buildings/7a9b3c1d-2e4f-4b6a-8c0d-1e2f3a4b5c6d"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProblemDetail": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "format": "uri"
                    },
                    "title": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "detail": {
                        "type": "string"
                    },
                    "instance": {
                        "type": "string",
                        "format": "uri"
                    },
                    "properties": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            },
            "BuildingRequest": {
                "type": "object",
                "description": "Data required to register a building",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Name of the building",
                        "example": "Torre Miraflores",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "address": {
                        "type": "string",
                        "description": "Street address of the building",
                        "example": "Av. Larco 1234, Miraflores",
                        "maxLength": 200,
                        "minLength": 0
                    },
                    "floors": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Number of floors",
                        "example": 12,
                        "minimum": 1
                    },
                    "constructionYear": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Year the building was built",
                        "example": 2015,
                        "maximum": 2100,
                        "minimum": 1800
                    },
                    "units": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Number of housing units",
                        "example": 48,
                        "minimum": 1
                    },
                    "administratorUserId": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Identifier of the administrator user who registers the building",
                        "example": "c1d2e3f4-a5b6-4c7d-8e9f-0a1b2c3d4e5f"
                    }
                },
                "required": [
                    "address",
                    "administratorUserId",
                    "constructionYear",
                    "floors",
                    "name",
                    "units"
                ]
            },
            "BuildingResponse": {
                "type": "object",
                "description": "A registered building",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Identifier generated by the service",
                        "example": "7a9b3c1d-2e4f-4b6a-8c0d-1e2f3a4b5c6d"
                    },
                    "name": {
                        "type": "string",
                        "description": "Name of the building",
                        "example": "Torre Miraflores"
                    },
                    "address": {
                        "type": "string",
                        "description": "Street address of the building",
                        "example": "Av. Larco 1234, Miraflores"
                    },
                    "floors": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Number of floors",
                        "example": 12
                    },
                    "constructionYear": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Year the building was built",
                        "example": 2015
                    },
                    "units": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Number of housing units",
                        "example": 48
                    },
                    "status": {
                        "type": "string",
                        "description": "PENDING_SENSORS until the first sensor is activated, then ACTIVE",
                        "enum": [
                            "PENDING_SENSORS",
                            "ACTIVE"
                        ],
                        "example": "PENDING_SENSORS"
                    },
                    "administratorUserId": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Identifier of the administrator user",
                        "example": "c1d2e3f4-a5b6-4c7d-8e9f-0a1b2c3d4e5f"
                    }
                }
            },
...
        }
    }
}