# src/main/java/com/rumi/buildingmanagement/infrastructure/web/dto/BuildingResponseDto.java
package com.rumi.buildingmanagement.infrastructure.web.dto;

import com.rumi.buildingmanagement.domain.model.BuildingStatus;
import io.swagger.v3.oas.annotations.media.Schema;

import java.util.UUID;

@Schema(name = "BuildingResponse", description = "A registered building")
public record BuildingResponseDto(
        @Schema(description = "Identifier generated by the service",
                example = "7a9b3c1d-2e4f-4b6a-8c0d-1e2f3a4b5c6d")
        UUID id,
        @Schema(description = "Name of the building", example = "Torre Miraflores")
        String name,
        @Schema(description = "Street address of the building", example = "Av. Larco 1234, Miraflores")
        String address,
        @Schema(description = "Number of floors", example = "12")
        int floors,
        @Schema(description = "Year the building was built", example = "2015")
        int constructionYear,
        @Schema(description = "Number of housing units", example = "48")
        int units,
        @Schema(description = "PENDING_SENSORS until the first sensor is activated, then ACTIVE",
                example = "PENDING_SENSORS")
        BuildingStatus status,
        @Schema(description = "Identifier of the administrator user",
                example = "c1d2e3f4-a5b6-4c7d-8e9f-0a1b2c3d4e5f")
        UUID administratorUserId
) {
}